モバイル用ヘッダーコンポーネント整理 #15
