+++ src/components/atoms/MobileSubMenuItem/index.jsx
@@ -1,0 +1,19 @@
+import {
+  ListItem,
+  ListItemButton,
+  ListItemIcon,
+  ListItemText,
+} from "@mui/material";
+
+const MobileSubMenuItem = ({ subIcon, subContent, route }) => {
+  return (
+    <ListItem>
+      <ListItemButton href={route}>
+        <ListItemIcon>{subIcon}</ListItemIcon>
+        <ListItemText primary={subContent} />
+      </ListItemButton>
+    </ListItem>
+  );
+};
+
+export default MobileSubMenuItem;
