--- src/components/layout/Header.jsx
@@ -1,12 +1,0 @@
-import MobileHeader from "../MobileHeader/MobileHeader";
-import TabletHeader from "../organisms/TabletHeader";
-
-export default function Header({ children }) {
-  return (
-    <>
-      <TabletHeader />
-      <MobileHeader />
-      {/* <main className="main">{children}</main> */}
-    </>
-  );
-}
